rename project to be more explicit
given that data-fusion has many conotations and different
meanings depending on the data being fused,
it makes sense to be more explicit in the intentions of this package.
The data-fusion methods here are also known as `statistical matching`
methods in data-fusion literature; particularly the literature focused
on marketing and the social sciences.
The data is typically panels, surveys or similarly measured record data.
Specifying `sm` in the package clarifies its intensions and use cases.