add predictive isotonic matching
Predictive Isotonic Matching is similar to Predictive Mean Matching
in that, first a model is built to predict the target for donors and
recipients, then records are matched on these predicted values.  However
, instead of matching on nearest value, matches are done based on
the rank relative to each data set. So, the rank is more important than
the value itself.